+++ Changelog.md
@@ -1,6 +1,10 @@
 # Changelog
 
 ## [0.1.5-dev][Unreleased] - Unreleased
+* Switch to the openwrt/sdk Docker image
+* Replace bind mounts for `sdk/staging_dir/hostpkg` and
+  `sdk/staging_dir/target` with one bind mount for `sdk/staging_dir`
+* Change bind mount destination paths for `packages` and `sdk/overrides`
 * Use `src-git-full` instead of `src-git` in feeds.conf
 
 ## [0.1.4] - 2022-05-30